+++ packages/renderscope-ui/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to the `renderscope-ui` package will be documented in this f
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2026-05-11
+
+First public release.
+
 ## [0.1.0] - 2026-02-19
 
 ### Added